requirements.txt
flask
requests
transbank-sdk